dashboard: reorg webpack and jest setup
After upgrading node deps to the latest, jest started failing and
took us down a rabbit hole. Took this opportunity to move webpack
and jest config files out of src to their own folders.

Eventually the issue was traced down to 'node-fetch' module which
breaks Jest because the latest is a pure ESM module and Jest still
depends on Babel to translate ESM to CJS...

The temporary fix is to downgrade 'node-fetch' down to '2.6.2'

NOTE: it is perhaps feasible to create a configuration to transpile
all dependencies of 'node-fetch' to CJS, instead of just the top
level of imports. But the effort is not worth it.